requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
fitz
wikipedia
requests
langchain-text-splitters
langchain-openai
chromadb
langchain-ollama
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq 
langchain_core
fastapi
uvicorn
langserve 
sse_starlette
duckduckgo-search
mysql-connector-python
SQLAlchemy
validators
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
cassio
datasets
openai 
tiktoken
gradio
boto3
awscli
langchain_nvidia_ai_endpoints
crewai
crewai_tools